+++ maple/build.gradle.kts
@@ -36,7 +36,7 @@ dependencies {
     implementation("com.fasterxml.jackson.dataformat:jackson-dataformat-yaml:2.17.1")
     implementation("com.fasterxml.jackson.module:jackson-module-kotlin:2.17.1")
     // kgit (replace with correct coords if different)
-    implementation("io.github.bitspittle:kgit:0.1.3") // <-- adjust to the lib you actually want
+//    implementation("io.github.bitspittle:kgit:0.1.3") // <-- adjust to the lib you actually want
     // ensure test deps for JUnit if not present
 }
