+++ maple/src/main/kotlin/com/tazzledazzle/maple/fixtures/BomEntry.kt
@@ -1,0 +1,9 @@
+package com.tazzledazzle.maple.fixtures
+
+data class BomEntry(
+    val name: String,
+    val version: String,
+    val repoUrl: String,
+    val firstParty: Boolean = true,
+    val buildCmd: String
+)
