+++ maple/src/main/kotlin/com/tazzledazzle/maple/fixtures/FixtureService.kt
@@ -61,11 +61,3 @@ class FixturesService(
     }
 }
 
-data class BomEntry(
-    val name: String,
-    val version: String,
-    val repoUrl: String,
-    val firstParty: Boolean = true,
-    val buildCmd: String
-)
-
